+++ __tests__/__integration__/__scripts__/imperative_ssh_delete_profile.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-set -e
-
-echo "================SSH DELETE PROFILE==============="
-zowe profiles delete ssh $*
-if [ $? -gt 0 ]
-then
-    exit $?
-fi