+++ jest.config.js
@@ -23,7 +23,8 @@ const sharedConfig = {
     "testRegex": "__tests__.*\\.(spec|test)\\.ts$",
     "moduleFileExtensions": [
         "ts",
-        "js"
+        "js",
+        "json"
     ],
     "testEnvironment": "node",
     "testPathIgnorePatterns": [
@@ -149,5 +150,7 @@ module.exports = {
         "!**/node_modules/**",
         "!**/lib/**"
     ],
-    "maxWorkers": "67%", // You may need to specify maxWorkers if you run out of RAM
+    // You may need to specify maxWorkers if you run out of RAM
+    // GHA should use 75% due to high ram, low core count, end user systems ~67%
+    "maxWorkers": process.env.GITHUB_ACTIONS != null ? "75%" : "67%",
 }